Set Models to the appropriate footprints (of their components)
Going Right-Click>Properties and checking the footprint under "Models" in the bottom-right showed they referred to non-existent footprints. I checked the "Link to Library Component" and noted the Model/footprint for the component. I then went and changed the Model to link to the aforementioned Model.
The offending parts seemed to be:
DA
H1
H2
H3
LRG1
S1
S2
S3